SSR (#7)
* Initial install

* Set up Scully configuration

* Add users route for http testing

* Add users route

* Add single users page, wire up some routing

* remove hash routing which fixes scully

* Merge AU with Scully

* Add conditional for scully

* Remove trailing slash from links

* Switch to web api for users gen

* Add support for remote webapi path list

* Add amp plugin, clean up api routes example

* Enforce trailing slash for urls

* Add remove FOUC plugin

* Clean up SSR build commands

* Enforce trailing slash with routing.

* Add in DOM asbtraction service, wire up to login

* Add app ID to env config

* Add transfer state

* Sync with main

* Rename fonts module to icons

* Various cleanup items

* Add users CRUD example

* Updates to user demo, fix ssr icons issue

* Add support for remote webapi path list

* Add amp plugin, clean up api routes example

* Enforce trailing slash for urls

* Add remove FOUC plugin

* Enforce trailing slash with routing.

* Merge branch 'main' into dev/ssr

# Conflicts:
#	package-lock.json
#	package.json
#	src/app/app.component.ts
#	src/app/app.module.ts
#	src/app/components/masterpage/nav/nav.component.html
#	src/app/components/masterpage/nav/nav.component.scss
#	src/app/components/masterpage/nav/nav.component.ts
#	src/app/vendor.module.ts

* adds the scully stuff back to app.module

* fixing merge errors

+++ src/app/shared/interceptors/error.interceptor.ts
@@ -2,6 +2,7 @@ import { ErrorHandler, Injectable } from '@angular/core';
 
 import { environment } from '$env';
 import { SettingsService } from '$settings';
+import { isBrowser } from '../services';
 
 interface AngularError {
   promise: any;
@@ -40,6 +41,10 @@ export class GlobalErrorHandler implements ErrorHandler {
       // this.settings.error$.next(error.message);
       this.resetState(error);
     }
+    // Display error in node
+    if (!this.settings.isBrowser) {
+      console.error(error);
+    }
     // Now throw the error to the console
     throw error;
   } // end handleError
@@ -49,6 +54,9 @@ export class GlobalErrorHandler implements ErrorHandler {
    * @param error
    */
   private resetState(error: AngularError) {
+    if (!isBrowser) {
+      return;
+    }
     console.error({ error: error });
     // const resetAction = () => {
     localStorage.clear();
